Se agregan los filtros de RoleAccessChecker a vista planes y se integra al menu default de usuario CPE

+++ controllers/PlanesController.php
@@ -12,6 +12,9 @@ use app\models\Instituto;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
+use app\commands\RoleAccessChecker;
+
 
 /**
  * PlanesController implements the CRUD actions for Planes model.
@@ -24,6 +27,27 @@ class PlanesController extends Controller
     public function behaviors()
     {
         return [
+              'access' => [
+                 'class' => AccessControl::className(),
+                 'only' => ['index', 'view', 'update', 'create', 'delete',],
+                 'rules' => [
+                     [
+                         'allow' => true,
+                         'actions' => ['',],
+                         'roles' => ['?'],
+                     ],
+                     [
+                         'allow' => true,
+                         'actions' => ['index', 'view', 'update', 'create', 'delete',],
+                         'roles' => ['@'],
+                     ],
+                     [
+                         'allow' => false,
+                         'actions' => ['',],
+                         'roles' => ['@'],
+                     ],
+                 ],
+             ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -39,13 +63,17 @@ class PlanesController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new PlanesSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+		if (RoleAccessChecker::actionIsAsignSector('planes/index')) {
+			try{
+				$searchModel = new PlanesSearch();
+				$dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+				return $this->render('index', [
+					'searchModel' => $searchModel,
+					'dataProvider' => $dataProvider,
+				]);
+			} catch (\yii\db\Exception $e) {return $this->redirect(['error/db-grant-error',]);}
+		} else return $this->redirect(['error/level-access-error',]);
     }
 
     /**
@@ -53,11 +81,14 @@ class PlanesController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionView($id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+    public function actionView($id){
+		if (RoleAccessChecker::actionIsAsignSector('planes/view')) {
+			try{
+				return $this->render('view', [
+					'model' => $this->findModel($id),
+				]);
+			} catch (\yii\db\Exception $e) {return $this->redirect(['error/db-grant-error',]);}
+		} else return $this->redirect(['error/level-access-error',]);
     }
 
     /**
@@ -65,25 +96,28 @@ class PlanesController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
-    public function actionCreate()
-    {
-        $model = new Planes();
-        $subModel = new Ano();
-        $subModel2 = new Materia();
-        $subModel3 = new Carrera();
-        $subModel4 = new Instituto();
+    public function actionCreate(){
+		if (RoleAccessChecker::actionIsAsignSector('planes/create')) {
+			try{
+				$model = new Planes();
+				$subModel = new Ano();
+				$subModel2 = new Materia();
+				$subModel3 = new Carrera();
+				$subModel4 = new Instituto();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->planes_id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-                'subModel' => $subModel,
-                'subModel2' => $subModel2,
-                'subModel3' => $subModel3,
-                'subModel4' => $subModel4,
-            ]);
-        }
+				if ($model->load(Yii::$app->request->post()) && $model->save()) {
+					return $this->redirect(['view', 'id' => $model->planes_id]);
+				} else {
+					return $this->render('create', [
+						'model' => $model,
+						'subModel' => $subModel,
+						'subModel2' => $subModel2,
+						'subModel3' => $subModel3,
+						'subModel4' => $subModel4,
+					]);
+				}
+			} catch (\yii\db\Exception $e) {return $this->redirect(['error/db-grant-error',]);}
+		} else return $this->redirect(['error/level-access-error',]);
     }
 
     /**
@@ -92,17 +126,20 @@ class PlanesController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionUpdate($id)
-    {
-        $model = $this->findModel($id);
+    public function actionUpdate($id){
+		if (RoleAccessChecker::actionIsAsignSector('planes/update')) {
+			try{
+				$model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->planes_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
-        }
+				if ($model->load(Yii::$app->request->post()) && $model->save()) {
+					return $this->redirect(['view', 'id' => $model->planes_id]);
+				} else {
+					return $this->render('update', [
+						'model' => $model,
+					]);
+				}
+			} catch (\yii\db\Exception $e) {return $this->redirect(['error/db-grant-error',]);}
+		} else return $this->redirect(['error/level-access-error',]);
     }
 
     /**
@@ -111,11 +148,13 @@ class PlanesController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+    public function actionDelete($id){
+		if (RoleAccessChecker::actionIsAsignSector('planes/delete')) {
+			try{
+				$this->findModel($id)->delete();
+				return $this->redirect(['index']);
+			} catch (\yii\db\Exception $e) {return $this->redirect(['error/db-grant-error',]);}
+		} else return $this->redirect(['error/level-access-error',]);
     }
 
     /**
